requirements.txt
 openai
wikipedia
faiss-cpu
python-dotenv